requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain_nvidia_ai_endpoints==0.3.10
gunicorn==21.2.0
python-dotenv==1.0.0
markdown2